change points to be based off damage
also tuned some values since proj weapons were 11x better xD

+++ Assets/Scripts/WaveSystem/StyleMeter/PointManager.cs
@@ -101,9 +101,9 @@ public class PointManager : MonoBehaviour
         UpdateUI();
     }
 
-    public void AddPoints(int amount)
+    public void AddPoints(float amount)
     {
-        int multipliedAmount = Mathf.RoundToInt(amount * currentMultiplier);
+        float multipliedAmount = Mathf.RoundToInt(amount * currentMultiplier);
         currentPoints += multipliedAmount;
         noDecayTimer = 0f;
         if (currentPoints > maxPoints) { currentPoints = maxPoints; }
